+++ tests/parsers/service.parser.spec.ts
@@ -170,6 +170,24 @@ describe('ServiceParser', () => {
 		expect(keys).to.deep.equal(['Hello', 'World']);
 	});
 
+	it('should NOT extract empty strings from the get()/instant()/stream() methods', () => {
+		const contents = `
+			@Component({ })
+			export class AppComponent {
+				public constructor(protected _translateService: TranslateService) { }
+				public test() {
+					this._translateService.get('');
+					this._translateService.instant('');
+					this._translateService.stream('');
+					this._translateService.get(['', '']);
+					this._translateService.instant(['', '']);
+					this._translateService.stream(['', '']);
+				}
+		`;
+		const keys = parser.extract(contents, componentFilename)?.keys();
+		expect(keys).to.deep.equal([]);
+	});
+
 	it('should not extract strings in get()/instant()/stream() methods of other services', () => {
 		const contents = `
 			@Component({ })
@@ -266,8 +284,9 @@ describe('ServiceParser', () => {
 	});
 
 	it('should extract strings from all classes in the file', () => {
+		// language=typescript
 		const contents = `
-			import { Injectable } from '@angular/core';
+			import { Injectable, inject } from '@angular/core';
 			import { TranslateService } from '@ngx-translate/core';
 			export class Stuff {
 				thing: string;
@@ -342,6 +361,44 @@ describe('ServiceParser', () => {
 		`;
 		const keys = parser.extract(contents, componentFilename)?.keys();
 		expect(keys).to.deep.equal(['Hello World']);
+	});
+
+	it('should extract strings when TranslateService is injected using the inject function within the constructor', () => {
+		const contents = `
+			export class MyComponent {
+				readonly propInstant: string;
+				readonly propStream: string;
+				readonly propGet: string;
+
+				constructor() {
+					const translate = inject(TranslateService);
+
+					this.propInstant = translate.instant('key.instant');
+					this.propStream = translate.stream('key.stream');
+					this.propGet = translate.get('key.get');
+				}
+			}
+		`;
+		const keys = parser.extract(contents, componentFilename)?.keys();
+		expect(keys).to.deep.equal(['key.instant', 'key.stream', 'key.get']);
+	});
+
+	it('should extract strings when TranslateService is injected using the inject function inline within the constructor', () => {
+		const contents = `
+			export class MyComponent {
+				readonly propInstant: string;
+				readonly propStream: string;
+				readonly propGet: string;
+
+				constructor() {
+					this.propInstant = inject(TranslateService).instant('key.instant');
+					this.propStream = inject(TranslateService).stream('key.stream');
+					this.propGet = inject(TranslateService).get('key.get');
+				}
+			}
+		`;
+		const keys = parser.extract(contents, componentFilename)?.keys();
+		expect(keys).to.deep.equal(['key.instant', 'key.stream', 'key.get']);
 	});
 
 	it('should locate TranslateService when it is a JavaScript native private property', () => {
@@ -796,6 +853,21 @@ describe('ServiceParser', () => {
 			`;
 			const keys = parser.extract(contents, componentFilename)?.keys();
 			expect(keys).to.deep.equal(['translation.key']);
+		});
+
+		it('should ignore injected services that are not TranslateService type', () => {
+			const contents = `
+				import {inject} from '@angular/core';
+				import {HttpClient} from '@angular/common/http';
+
+				export const getResource = (): string => {
+					const someService = inject(SomeService).instant('property');
+					const someStream = inject(SomeStreamService).stream('property');
+					return inject(HttpClient).get('/path/to/resource/');
+				}
+			`;
+			const keys = parser.extract(contents, componentFilename)?.keys();
+			expect(keys).to.deep.equal([]);
 		});
 	});
 });